add get token files and rename extractor file
